# docs/openapi-ccn.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "3.2.4-ccn.4",
    "title": "DocuSeal CCN API (fork additions)",
    "description": "Operations the CCN fork adds to DocuSeal's API (docs/openapi.json documents the upstream ones, all implemented). Administration functions that are UI-only in DocuSeal (users, webhooks, account settings, template folders, template versions, field detection) plus the `preferences` extension of PUT /templates/{id}. Same authentication as upstream (X-Auth-Token). Every refused request answers 422 {\"error\": \"...\"}.",
    "contact": {
      "name": "docuseal-ccn",
      "url": "https://github.com/lucouto/docuseal-ccn"
    }
  },
  "servers": [
    {
      "url": "https://docuseal.cheminneuf.community/api",
      "description": "Production (stock 3.2.4 until promoted)"
    },
    {
      "url": "https://docuseal-staging.cheminneuf.community/api",
      "description": "Staging (the fork)"
    }
  ],
...
  "paths": {
...
    "/ccn/templates/{id}/detect_fields": {
      "post": {
        "security": [
          {
            "AuthToken": []
          }
        ],
        "tags": [
          "Templates"
        ],
        "summary": "Detect fields",
        "description": "Runs the field detector on the template's documents (or one document, one page) and returns the candidates; with `apply` they are added to the first signer, skipping candidates that overlap an existing field. At most 30 pages per request.",
        "operationId": "detectTemplateFields",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The unique identifier of the template.",
            "example": 1000001
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "attachment_uuid": {
                    "type": "string",
                    "description": "Only this document."
                  },
                  "page": {
                    "type": "integer",
                    "description": "Only this page (1-based)."
                  },
                  "apply": {
                    "type": "boolean",
                    "description": "Add the candidates to the template (default false).",
                    "default": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "documents": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "attachment_uuid": {
                            "type": "string"
                          },
                          "pages": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "page": {
                                  "type": "integer",
                                  "description": "1-based page number."
                                },
                                "fields": {
                                  "type": "array",
                                  "items": {
                                    "type": "object"
                                  },
                                  "description": "Candidate fields in the template's stored shape (areas[].page is 0-based)."
                                }
                              },
                              "required": [
                                "page",
                                "fields"
                              ]
                            }
                          }
                        },
                        "required": [
                          "attachment_uuid",
                          "pages"
                        ]
                      }
                    },
                    "applied": {
                      "type": "integer",
                      "description": "Number of candidates added to the template (0 unless apply is true)."
                    }
                  },
                  "required": [
                    "documents",
                    "applied"
                  ]
                },
                "example": {
                  "documents": [
                    {
                      "attachment_uuid": "c1a2...",
                      "pages": [
                        {
                          "page": 1,
                          "fields": [
                            {
                              "uuid": "...",
                              "type": "text",
                              "required": false,
                              "areas": [
                                {
                                  "x": 0.1,
                                  "y": 0.2,
                                  "w": 0.3,
                                  "h": 0.04,
                                  "page": 0,
                                  "attachment_uuid": "c1a2..."
                                }
                              ]
                            }
                          ]
                        }
                      ]
                    }
                  ],
                  "applied": 0
                }
              }
            }
          },
          "422": {
            "description": "The request was refused; `error` says why.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                },
                "example": {
                  "error": "Unknown setting 'nope'"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AuthToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Auth-Token"
      }
    }
  }
}